# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Voetbalclub Kantine Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /volunteers:
    get:
      operationId: listVolunteers
      tags: [volunteers]
      summary: List all volunteers
      responses:
        "200":
          description: List of volunteers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Volunteer"
    post:
      operationId: createVolunteer
      tags: [volunteers]
      summary: Create a volunteer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVolunteerRequest"
      responses:
        "201":
          description: Volunteer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volunteer"
  /volunteers/{id}:
    get:
      operationId: getVolunteer
      tags: [volunteers]
      summary: Get a volunteer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Volunteer found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volunteer"
        "404":
          description: Not found
    put:
      operationId: updateVolunteer
      tags: [volunteers]
      summary: Update a volunteer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVolunteerRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volunteer"
    delete:
      operationId: deleteVolunteer
      tags: [volunteers]
      summary: Delete a volunteer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
components:
  schemas:
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    Volunteer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        hasPassword:
          type: boolean
          description: "True als er een wachtwoord is ingesteld voor dit account"
        isAdmin:
          type: boolean
          description: "True als deze vrijwilliger beheerdersrechten heeft"
        availability:
          type: array
          items:
            $ref: "#/components/schemas/ShiftSlot"
        groupId:
          type: integer
          nullable: true
        groupMembers:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
            required: [id, name]
        createdAt:
          type: string
      required: [id, name, availability, groupMembers, createdAt]
    CreateVolunteerRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
          description: "Wachtwoord voor vrijwilliger login (wordt gehasht opgeslagen)"
        isAdmin:
          type: boolean
          nullable: true
          description: "Geeft de vrijwilliger beheerdersrechten"
        availability:
          type: array
          items:
            $ref: "#/components/schemas/ShiftSlot"
        groupMemberIds:
          type: array
          nullable: true
          items:
            type: integer
      required: [name]
    ShiftSlot:
      type: string
...